code review
Good job on your project! All the functions are clear and easy to find, and I like that you neatly separate file in different folders so it's clearly organized. As a small advice, it's a good idea to also organize all your css files in a separate folder so they are easier to find if other users want to see or modify them. Other than that, great job!