Review by vicki
I really like how you implemented this page. It is very straight forward and learned from it. I really like the design of the project. It does very well in having a uniform  color palette. I was able to use tab on a ll pages. I really like the create a recipe page as all feature are easy to use and allows user creativity to add any ingredients they want along with a picture. When login in it would be nice to a feature that allows you see and hide password when typing. This can help if a user getting their password wrong but maybe 1 letter off. Overall code is well structure, organized, and clean. I really like the idea of your project! It has good User Interface and all features are straight forward to use.